Make the init script able to source in a shell file if present.
This allows for a configuration file, /etc/gitlab.conf
by default be defined with any site specific values to be set
so sites which need to differ from the defaults can override
as appropriate.

Example would be /etc/gitlab.conf:
app_root="/some/dir/with/gitlab/in/it"

This means no links or other hacks would be necessary for
site specific customization.